"Fix" for 'upload_motion_to_dropbox' Celery task's concurrency: The 'lock' mechanism was removed, the task is now not thread-safe, meaning it shouldn't be dispatched more than once at the same time.
To prevent it from happening, the Celery worker should be allowed to run only one task concurrently, like so:
> python manage.py celery worker -l info --concurency=1

This will also make other celery tasks run only when others are not running,
but it's not critical for the current system implementation.
If other tasks are added, another solution should be considered.
A simple solution is to run another worker for other tasks by separating the tasks to different Celery apps.